♻️ Refactor: Mapper 클래스 분리
