+++ src/main/java/com/likelion/backendplus4/yakplus/drug/application/service/DrugSymptomService.java
@@ -10,12 +10,17 @@ import com.likelion.backendplus4.yakplus.drug.domain.model.GovDrug;
 import com.likelion.backendplus4.yakplus.drug.infrastructure.adapter.DrugSymptomEsAdapter;
 import com.likelion.backendplus4.yakplus.drug.infrastructure.adapter.GovDrugJpaAdapter;
 import com.likelion.backendplus4.yakplus.drug.infrastructure.adapter.persistence.repository.document.DrugSymptomDocument;
-import com.likelion.backendplus4.yakplus.drug.infrastructure.support.mapper.DrugDataMapper;
+import com.likelion.backendplus4.yakplus.drug.infrastructure.support.mapper.EntityDocMapper;
 import com.likelion.backendplus4.yakplus.drug.presentation.controller.dto.DrugSymptomSearchListResponse;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+/**
+ * 약품 증상 데이터를 처리하는 서비스입니다.
+ * @sice 2025-04-24
+ * @modified 2025-04-25
+ */
 @Service
 @RequiredArgsConstructor
 @Slf4j
@@ -26,6 +31,14 @@ public class DrugSymptomService {
 	private final GovDrugJpaAdapter drugJpaAdapter;
 	private final DrugSymptomEsAdapter symptomAdapter;
 
+    /**
+     * DB에서 약품 데이터를 페이징으로 가져와 Elasticsearch에 일괄 색인합니다.
+     * 각 페이지는 CHUNK_SIZE만큼 처리되며, 모든 데이터를 순차적으로 색인합니다.
+     *
+     * @author 박찬병
+     * @since 2025-04-24
+     * @modified 2025-04-25
+     */
 	public void indexAll() {
 		int page = 0;
 		Page<GovDrug> drugPage;
@@ -36,7 +49,7 @@ public class DrugSymptomService {
 
 			// 2. 도메인 → ES Document 변환
 			List<DrugSymptomDocument> docs = drugPage.stream()
-				.map(DrugDataMapper::toDocument)  // 내부에서 예외 처리 됨
+				.map(EntityDocMapper::toDocument)  // 내부에서 예외 처리 됨
 				.toList();
 
 			// 3. 청크별 ES에 색인
@@ -47,7 +60,17 @@ public class DrugSymptomService {
 		} while (drugPage.hasNext());
 	}
 
-	public DrugSymptomSearchListResponse getSymptomAutoComplete(String q) {
-		return new DrugSymptomSearchListResponse(symptomAdapter.getSearchAutoCompleteResponse(q));
-	}
+    /**
+     * 주어진 사용자 입력 문자열을 바탕으로 증상 자동완성 키워드를 가져옵니다.
+     * Elasticsearch에서 Suggest API 등을 활용하여 추천 결과를 반환합니다.
+     *
+     * @param q 사용자 입력 문자열
+     * @return 자동완성 추천 결과 리스트 DTO
+     * @author 박찬병
+     * @since 2025-04-24
+     * @modified 2025-04-25
+     */
+    public DrugSymptomSearchListResponse getSymptomAutoComplete(String q) {
+        return new DrugSymptomSearchListResponse(symptomAdapter.getSearchAutoCompleteResponse(q));
+    }
 }
