+++ src/main/java/com/likelion/backendplus4/yakplus/drug/infrastructure/support/mapper/DrugDataMapper.java
@@ -1,15 +1,7 @@
 package com.likelion.backendplus4.yakplus.drug.infrastructure.support.mapper;
 
-import java.io.IOException;
-import java.util.List;
-
-import com.likelion.backendplus4.yakplus.drug.exception.ScraperException;
-import com.likelion.backendplus4.yakplus.drug.exception.error.ScraperErrorCode;
-import com.likelion.backendplus4.yakplus.drug.infrastructure.adapter.persistence.repository.document.DrugSymptomDocument;
 import com.likelion.backendplus4.yakplus.drug.infrastructure.adapter.persistence.repository.entity.GovDrugEntity;
 import com.likelion.backendplus4.yakplus.drug.domain.model.GovDrug;
-import com.likelion.backendplus4.yakplus.drug.infrastructure.support.parser.SymptomTextParser;
-import com.likelion.backendplus4.yakplus.drug.infrastructure.support.parser.JsonTextParser;
 
 public class DrugDataMapper {
 	public static GovDrug toDomainFromEntity(GovDrugEntity e) {
@@ -29,26 +21,5 @@ public class DrugDataMapper {
 			.build();
 	}
 
-	/**
-	 * DB 엔티티 → ES Document 변환
-	 */
-	public static DrugSymptomDocument toDocument(GovDrug entity) {
-		List<String> raws;
-		try {
-			raws = JsonTextParser.extractAllTexts(entity.getEfficacy());
-		} catch (IOException e) {
-			// 파싱 실패 시 커스텀 예외 던짐
-			throw new ScraperException(ScraperErrorCode.PARSING_ERROR);
-		}
 
-		String flatText = SymptomTextParser.flattenLines(raws);
-		List<String> suggestTokens = SymptomTextParser.tokenizeForSuggestion(flatText);
-
-		return DrugSymptomDocument.builder()
-			.drugId(entity.getDrugId())
-			.drugName(entity.getDrugName())
-			.symptom(flatText)
-			.symptomSuggester(suggestTokens)
-			.build();
-	}
 }
